refactor: organize files and remove extras
